+++ pydidas/logging_level.py
@@ -16,8 +16,7 @@
 # along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
 
 """
-The core package defines base classes used throughout the full pydidas
-suite.
+The logging_level module sets the global logging level in pydidas.
 """
 
 __author__ = "Malte Storm"
@@ -25,36 +24,30 @@ __copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
-__all__ = []
+__all__ = ["LOGGING_LEVEL"]
 
 
-import os as __os
-import sys as __sys
-
-if "--qt6" in __sys.argv:
-    __os.environ["QT_API"] = "pyside6"
-    __os.environ["FORCE_QT_API"] = "True"
+import logging
+import re
+import sys
 
 
-import qtpy
+__log_keys = ["logging", "logger", "logging-level", "log"]
+__log_levels = ["DEBUG", "INFO", "WARN" "WARNING", "ERROR", "CRITICAL", "FATAL"]
+__log_level = logging.ERROR
 
-if "--verbose-qt" in __sys.argv:
-    print("pydidas Using QT: ", qtpy.API_NAME)
+
+if any(
+    re.match(rf"^-{{1,2}}{re.escape(key)}$", arg)
+    for key in __log_keys
+    for arg in sys.argv
+):
+    for _index, _arg in enumerate(sys.argv):
+        if _arg.strip("-") in __log_keys and _index < len(sys.argv) - 1:
+            _level = sys.argv[_index + 1].upper()
+            if _level in __log_levels:
+                __log_level = getattr(logging, _level)
+                break
 
 
-# import __all__ items from modules:
-from .fontsize import *
-from .pydidas_qapp import *
-from .pydidas_splash_screen import *
-
-# add modules' __all__ items to package's __all__ items and unclutter the
-# namespace by deleting the module references:
-from . import pydidas_qapp
-
-__all__.extend(pydidas_qapp.__all__)
-del pydidas_qapp
-
-from . import pydidas_splash_screen
-
-__all__.extend(pydidas_splash_screen.__all__)
-del pydidas_splash_screen
+LOGGING_LEVEL = __log_level